Erased unnecesary files and texts
