+++ src/components/Calculator.jsx
@@ -1,29 +1,6 @@
 import React from 'react';
-import PropTypes from 'prop-types';
+import Button from './Button';
 import './Calculator.css';
-
-function Button(props) {
-  const { value } = props;
-
-  // Define button types based on their values
-  let type = '';
-  if (['+', 'x', '-', '÷'].includes(value)) {
-    type = 'operation';
-  } else if (!Number.isNaN(Number(value))) {
-    type = 'number';
-  } else {
-    type = 'function';
-  }
-
-  return <button type="button" data-type={type} data-value={value}>{value}</button>;
-}
-
-Button.propTypes = {
-  value: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-  ]).isRequired,
-};
 
 function Calculator() {
   const buttons = ['AC', '+/-', '%', '÷', 7, 8, 9, 'x', 4, 5, 6, '-', 1, 2, 3, '+', 0, '.', '='];
